Make genderize work with active model validations
Make Gender a kind of String and don't raise exceptions with invalid
values so that gender can be validated using active model validations.

This is about the least invasive commit I could manage that would allow
genderize to work with built in validations in rails, without breaking
existing functionality and tests. It's far from ideal as it doesn't deal
with invalid values well and still modifies the string representing a
gender rather than just wrapping it add gender behaviour.